move download spec
